Add scene particle object
A few major downside to the built-in particle system was making progress difficult. This is a new custom particle implementation that solves these issues. This is not meant to be a total replacement for the main particle system, but a parallel implementation when required. It's certainly going to be much slower that default CPUParticles, for instance.
1. Cannot spawn a scene as a particle
2. No Brownian motion
3. Need to dynamicly modify particle forces based on scene