Add world command
This allows worlds to be loaded via console command.
It will first search for worlds internally, then externally.
Example:
`world "Nifflas - The Machine.knytt.bin"`
(Make sure to use quotes if there are spaces in the name)

It also take an optional boolean that tells whether or not to play the
intro.

This is in preparation to allow better automated testing scripts going
forward.